Serve blockchain analytics data from the new Flink pipeline
Removed all the code around fetching the caching the data from the old
Etherbi into influxDB. The data will come directly from influxDB from
now on and will be kept up-to-date by Flink pipelines.
