+++ config/config.exs
@@ -58,22 +58,22 @@ config :sanbase, Sanbase.ExternalServices.TwitterData.Store,
   database: "twitter_followers_data"
 
 config :sanbase, Sanbase.Etherbi.Transactions.Store,
-  host: {:system, "INFLUXDB_HOST", "localhost"},
-  port: {:system, "INFLUXDB_PORT", 8086},
+  host: {:system, "ETHERBI_INFLUXDB_HOST", "localhost"},
+  port: {:system, "ETHERBI_INFLUXDB_HOST", 8086},
   pool: [max_overflow: 10, size: 20],
-  database: "transactions"
+  database: "erc20_exchange_transactions"
 
 config :sanbase, Sanbase.Etherbi.BurnRate.Store,
-  host: {:system, "INFLUXDB_HOST", "localhost"},
-  port: {:system, "INFLUXDB_PORT", 8086},
+  host: {:system, "ETHERBI_INFLUXDB_HOST", "localhost"},
+  port: {:system, "ETHERBI_INFLUXDB_HOST", 8086},
   pool: [max_overflow: 10, size: 20],
-  database: "burn_rate"
+  database: "erc20_burn_rate"
 
 config :sanbase, Sanbase.Etherbi.TransactionVolume.Store,
-  host: {:system, "INFLUXDB_HOST", "localhost"},
-  port: {:system, "INFLUXDB_PORT", 8086},
+  host: {:system, "ETHERBI_INFLUXDB_HOST", "localhost"},
+  port: {:system, "ETHERBI_INFLUXDB_HOST", 8086},
   pool: [max_overflow: 10, size: 20],
-  database: "transaction_volume"
+  database: "erc20_transaction_volume"
 
 config :sanbase, Sanbase.ExternalServices.Etherscan.Store,
   host: {:system, "INFLUXDB_HOST", "localhost"},
@@ -152,18 +152,6 @@ config :sanbase, Sanbase.ExternalServices.TwitterData.HistoricalData,
   # 1 day
   update_interval: 1000 * 60 * 60 * 24,
   sync_enabled: {:system, "TWITTERCOUNTER_SCRAPER_ENABLED", false}
-
-config :sanbase, Sanbase.Etherbi.Transactions,
-  update_interval: 1000 * 60 * 5,
-  sync_enabled: {:system, "ETHERBI_TRANSACTIONS_SYNC_ENABLED", false}
-
-config :sanbase, Sanbase.Etherbi.BurnRate,
-  update_interval: 1000 * 60 * 5,
-  sync_enabled: {:system, "ETHERBI_BURN_RATE_SYNC_ENABLED", false}
-
-config :sanbase, Sanbase.Etherbi.TransactionVolume,
-  update_interval: 1000 * 60 * 5,
-  sync_enabled: {:system, "ETHERBI_TRANSACTION_VOLUME_SYNC_ENABLED", false}
 
 config :sanbase, Sanbase.Notifications.CheckPrices,
   webhook_url: {:system, "CHECK_PRICES_WEBHOOK_URL"},
