+++ test/sanbase_web/graphql/etherbi_burn_rate_api_test.exs
@@ -8,8 +8,6 @@ defmodule Sanbase.Etherbi.BurnRateApiTest do
   import SanbaseWeb.Graphql.TestHelpers
 
   setup do
-    Application.put_env(:sanbase, Sanbase.Etherbi, use_cache: true)
-
     Store.create_db()
 
     ticker = "SAN"
